refactor: split files by functionality
