+++ src/main/kotlin/racingcar/Round.kt
@@ -1,0 +1,26 @@
+package racingcar
+
+data class Round(val time: Int, val cars: List<Car>) : Iterable<Car> {
+    override fun iterator(): Iterator<Car> = cars.iterator()
+    fun tryMoveForward(numbers: List<Int>): Round {
+        require(numbers.size == cars.size) { "숫자 수와 자동차 수가 동일해야합니다." }
+        return Round(time + 1, cars.zip(numbers).map { (car, number) ->
+            car.tryMoveForward(number)
+        })
+    }
+
+    private fun getMaxPosition() = (cars.maxBy { it.position }).position
+    fun getWinners(): List<Car> = cars.filter { getMaxPosition() == it.position }
+    fun totalCars(): Int = cars.size
+
+    companion object {
+        fun startWith(names: List<String>): Round {
+            val usedNames = mutableSetOf<String>()
+            return Round(0, names.map { name ->
+                require(!usedNames.contains(name)) { "중복된 자동차 이름($name)은 허용되지 않습니다." }
+                usedNames.add(name)
+                Car.withStartPosition(name)
+            })
+        }
+    }
+}
